Collapse equal terms.
Creating a tree with many duplicate elements causes the tree to become
degenerate (structure A -> A -> A -> A), which is memory and time
inefficient when the tree needs to query or add to this structure. We
can collapse terms equal to the term of the current node into that node
to avoid this. By the triangle inequality, any node equal to the current
node will meet the threshold if the current node meets the threshold as
well.

Making this change speeds up insertion and querying of duplicate entries
as well as significantly reducing memory footprint.